mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.32.0",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
